requirements.txt
# GUI 框架
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# UI 美化
PyQt6-Fluent-Widgets
pywin32; sys_platform == 'win32'

# 数据库 ORM
SQLAlchemy==2.0.23

# 配置管理
pydantic==2.5.2
pydantic-settings==2.1.0

# 数据处理
pandas==2.1.4
openpyxl==3.1.2

# 文档生成
python-docx==1.1.0
Pillow==10.1.0

# AI 与网络
openai==1.6.1
requests==2.31.0
urllib3==2.1.0

# 日志系统
loguru==0.7.2

# 打包工具（开发依赖）
pyinstaller==6.3.0
